Fix UFW_PLUGIN_HELPER_PATH if LIBEXEC_INSTALL_DIR is absolute
LIBEXEC_INSTALL_DIR can be either an absolute or a relative path. (CMake
accepts both.) If LIBEXEC_INSTALL_DIR is an absolute path (which it is
by default with the Fedora build macros, for technical reasons), the
path would end up as "/usr//usr/libexec/nomad_ufw_plugin_helper.py",
which is of course not valid. This commit fixes that.

Instead of attempting to compute the UFW_PLUGIN_HELPER_PATH within
ufw_helper_config.h.cmake, make it a CMake variable, and compute it in
CMakeLists.txt, checking for IS_ABSOLUTE.