Fix netstat usage
Netstat is a run time dependency, not a compile time one.
If one compiled the application without having netstat installed,
then even after installing it, the firewall wouldnt run.

It's better to just assume that the app is installed in the PATH
and give a warning if it isn't, asking for the user to install it